+++ src/core/windowagentbase.cpp
@@ -3,6 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "windowagentbase.h"
+#include "qwindowkit_linux.h"
 #include "windowagentbase_p.h"
 
 #include <QWKCore/qwkconfig.h>
@@ -15,7 +16,6 @@
 #  include "cocoawindowcontext_p.h"
 #else
 #  include "qtwindowcontext_p.h"
-#  include "linuxdesktopenvapi.h"
 #  include "linuxwaylandcontext_p.h"
 #  include "linuxx11context_p.h"
 #endif
@@ -59,12 +59,12 @@ namespace QWK {
 #elif defined(Q_OS_MAC) && !QWINDOWKIT_CONFIG(ENABLE_QT_WINDOW_CONTEXT)
         return new CocoaWindowContext();
 #else
-        if (LinuxDesktopEnvAPI::isX11Platform()) {
-            if (LinuxDesktopEnvAPI::x11API().isValid()) {
+        if (QWK::Private::isX11Platform()) {
+            if (QWK::Private::x11API().isValid()) {
                 return new LinuxX11Context();
             }
-        } else if (LinuxDesktopEnvAPI::isWaylandPlatform()) {
-            if (LinuxDesktopEnvAPI::waylandAPI().isValid()) {
+        } else if (QWK::Private::isWaylandPlatform()) {
+            if (QWK::Private::waylandAPI().isValid()) {
                 return new LinuxWaylandContext();
             }
         }
